+++ packages/arbitration/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @claudexor/arbitration
 
+## 3.2.1
+
+### Patch Changes
+
+- @claudexor/schema@3.2.1
+
 ## 3.2.0
 
 ### Patch Changes