+++ packages/gateway/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @claudexor/gateway
 
+## 3.2.1
+
+### Patch Changes
+
+- @claudexor/core@3.2.1
+- @claudexor/schema@3.2.1
+
 ## 3.2.0
 
 ### Patch Changes