+++ packages/journal/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @claudexor/journal
 
+## 3.2.1
+
+### Patch Changes
+
+- @claudexor/util@3.2.1
+
 ## 3.2.0
 
 ### Patch Changes