+++ packages/synthesis/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @claudexor/synthesis
 
+## 3.2.1
+
+### Patch Changes
+
+- @claudexor/arbitration@3.2.1
+
 ## 3.2.0
 
 ### Patch Changes